Minor changes in the getter methods for arrivalTime and centralFrequency
The arrivalTime is calculated, in the frequency domain, as  the center of mass of the groupDelay.
Before, the getter for arrivalTime did not return exactly the center of mass, but rather the value of the closest array element.
Now this is fixed.

A similar change has been done for the getter for the centralFrequency property.