extended functionality of increaseTimeResolution method
Now increaseTimeResolution allows to set the minumum number of points within an optical period
or within the pulse duration or within the pulse standard deviation or to set the total number of points.

Fixed a bug in the call to the zeropad function, which now receives the number of added points
instead of the new total number of points.

Once the new number of points has been estimated, it is rounded up to the closest even integer.
Before it was rounded up to the closest power of 2, instead.